pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tkdp-toolbox"
version = "0.2.0"
description = "Swiss army knife toolbox for a devops engineer."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.6"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
tkdp-toolbox = "tkdp_toolbox.__main__:main"
tkdptoolbox = "tkdp_toolbox.__main__:main"

[project.urls]
Homepage = "https://github.com/tkdpython/tkdp-toolbox"
